Add simplistic writer loop
Assumes a single process handling the topic + table. Does not implement
proper handling of Kafka offsets at all (relies solely on consumer group
offsets).
