+++ src/lib.rs
@@ -1,0 +1,4 @@
+mod deltalake_ext;
+mod kafka_delta_ingest;
+
+pub use kafka_delta_ingest::*;
